+++ registry/vue/ui/breadcrumb/Breadcrumb.vue
@@ -8,7 +8,7 @@ const props = defineProps<{
 
 <template>
   <nav
-    aria-label="breadcrumb"
+    aria-label="Fil d’Ariane"
     data-slot="breadcrumb"
     :class="props.class"
   >