+++ registry/vue/ui/sidebar/Sidebar.vue
@@ -42,8 +42,8 @@ const { isMobile, state, openMobile, setOpenMobile } = useSidebar()
       }"
     >
       <SheetHeader class="sr-only">
-        <SheetTitle>Sidebar</SheetTitle>
-        <SheetDescription>Displays the mobile sidebar.</SheetDescription>
+        <SheetTitle>Navigation</SheetTitle>
+        <SheetDescription>Menu de navigation principal.</SheetDescription>
       </SheetHeader>
       <div class="flex h-full w-full flex-col">
         <slot />